move libwhiscy as modules
