+++ pyproject.toml
@@ -28,6 +28,9 @@ whiscy_haddock = "whiscy.cli_haddock:main"
 include-package-data = true
 packages = ["src"]
 
+[tool.pytest.ini_options]
+pythonpath = ["src"]
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
