+++ src/whiscy/modules/quotes.py
